requirements.txt
openai
python-dotenv
sounddevice
numpy
faster-whisper
pydub
scipy
pyttsx3
transformers
torchaudio
librosa
soundfile
torch